Get rid of init methods and instead do it in the constructor. Use value types or smart pointers where raw pointers not necessary.

+++ sdlgame_win/src/ECS/TransformComponent.cpp
@@ -8,7 +8,8 @@ void TransformComponent::draw() {
 	}
 	auto relPos = Globals::get().cameraRelative(getPosition());
 	SDL_Rect transform_rect{ static_cast<int>(relPos.x), static_cast<int>(relPos.y),
-							 getWidth(), getHeight() };
+							 static_cast<int>(getWidth()),
+							 static_cast<int>(getHeight()) };
 	Utils::drawRect(&transform_rect, RGBVals::red());
 }
